Implement binary search to find users by username
Create a `binary_search` subroutine that will allow us to search through
an array.

Used Wikipedia and Rosetta Code as a reference:
  - https://en.wikipedia.org/wiki/Binary_search_algorithm#Algorithm
  - http://rosettacode.org/wiki/Binary_search#Perl

In order to test matches, `binary_search` expects an anonymous function
that will return 0, 1, or -1 depending on the result of the comparison.

The anonymous function we use compares usernames alphabetically to the
one we want to find.

Finally, we have a section commented out that performs a linear search
for a user in the members array. We will time this and the binary search
version to see how they compare. Slack attests that the `users.list` API
method provides users "in no particular order"
(https://api.slack.com/methods/users.list), however most users appear to
be ordered by username. The one exception appears to be 'slackbot', who
appears at the end of the list.