Add 'swhois' Irssi command
Create a command that will retrieve 'whois' information about a given
user. This command takes a username argument and searches our users.list
database for a match. It then prints out a field from the user hash.

We'll want to print out more information in a more structured display,
but this is just for testing within Irssi.

The server conditional check is commented out so I can test this without
connecting to a server.

Running `/swhois` without an argument should print whois information for
the current user/nick. Will have to look up how to get that information.