swhois: Don't print fields if their value is empty
Certain fields that we weren't checking against could be empty (title
and email). If that's the case, I don't think we should be printing a
line with a field label and empty value.

Add a new subroutine that prints the field if it's not blank. This
eliminates all the repetitive `if` statements.