If the users list is not available, fetch it from the Slack API
In `find_user`, we check whether the Storable file is there (and
non-zero). If it isn't, we fetch the users list from the Slack API.

Renamed the subroutine containing the HTTP code to `fetch_users_list`
from `users_list` to better convey the fact that it has to reach out to
the web in order to get it.

Store the Storable object in the Irssi scripts directory (instead of our
local development directory as it was previously).

Use the correct Slack API URL in our HTTP request, and also include our
Slack API token from the Irssi settings.

Add better error handling and messaging when the API request fails for
whatever reason. Die on error so we don't end up executing the
`retrieve` line (otherwise this line will cause an error complaining
that the `users.list.plstore` file doesn't exist if it hasn't yet been
created).

If the API call succeeds and Slack gives us an 'ok' response, then
`store` the JSON-decoded users list array in our serialized cache file
and assign it to the `@users_list` global variable so we additionally
have it in memory..