Only check for server if it's needed
We only need the `$server` object if we're doing a WHOIS for the current
nick. That's because if there's no server, there's no current nick.
Otherwise, as long as the Slack API token is configured, a server isn't
needed and shouldn't necessarily be required in order to request the
information.