Get & print user profile information from Slack
Fetch profile information from the Slack API's `users.profile.get`
method. Grab the custom fields from the response and print these as part
of the SWHOIS output.

Note that this currently only works when a nick argument is passed to
the command.

This is just an initial version. It should be cleaned up such that
getting the custom fields is probably in its own subroutine, and more
importantly that fetching the profile fields from the API doesn't stop
WHOIS output if the API call fails. After all, most WHOIS information
can be gathered from the local cache.