slack_api: Move `die`s outside the subroutine
Don't cause the command to break WHOIS output. Previously when we were
only using it to call the 'users.list' Slack API method it was fine
because the WHOIS output didn't directly depend on fetching the entire
users list (since we have a cache).

Now that we query the Slack API on each `/swhois` (the
'users.profile.get' method), we don't want to break profile output that
can easily come from our cached data by dying. There's still value in
displaying profile information without custom fields.

We remove the `die` calls from `slack_api` and change them to a single
`die` in the `fetch_users_list` method, since if this fails we really do
want to stop script execution. In that case, if we don't get the users
list back from the API, we can't set it to our in-memory cache nor store
it to our cache file.