Start of a subroutine to update profile fields
This doesn't work, but includes some rough code to allow updating of
profile fields.

A key and value are passed to the sub. These will be sent to the API as
the field to update. We get the user by querying with the current nick
(we'll have to get the `$server` from somewhere). Then a sample API call
using `slack_api`.

Also include some ideas on how we can update custom fields more easily
because their non-human-readable ids must be used as a key in order to
update them.