update_user_profile: Allow updating of custom fields
Slack provides a set of default fields on a profile (relevant ones for
editing are added to `@profile_fields`). Additionally, teams can add
custom profile fields. These use unique ids that must be pulled from a
profile object.

This change maps an underscorized field name with a custom field id so
that the id can be passed in the API call to change the field's value.

The `@profile_fields` array is for now just to remind me what fields are
available to change by default on profile. In the future, this will be
used for key name completion.