Add `/slack_profile_set` command
A new command that leverages `update_user_profile` to change a given
profile field.

Should this command delegate directly to `update_user_profile` (now that
I think about it) instead of using an intermediary subroutine? I'm not
sure.

Also note that we removed the `/slack_profile` command and changed
`/slack_profile sync` to `/slack_profile_sync`. I had trouble getting
the subcommand setup to work. When I tried adding a `/slack_profile set`
command, the `/slack_profile` command would be run when I tried to
invoke it. I tried `\&cmd_set` syntax, `sub { cmd_set(); }` syntax, etc.
for all the commands but it kept doing the same thing. Tried looking at
some other scripts to see how they do it, but couldn't really figure it
out except for 'trackbar' which uses a single subroutine for all it's
subcommands which I didn't want to do.

That's when I realised that subcommands kind of suck. I mean, you have
to tab complete twice in order to invoke the command. Wouldn't it be so
much nicer if you only had to tab complete the first command? So I
changed the spaces to underscores which I think is better.