Fix users list bug after running 'sync' command
After running the `/slack_profile_sync` command, you would get an error
trying to get the data from `@users_list` in the loop in `find_user`.

This had to do with retrieving the data from the serialised `Storable`
cache file.

I'm not sure why there's a difference between the `retrieve`d cache
version and the version downloaded and JSON decoded directly from the
API, but that discrepancy caused the error.

Now only refresh the `@users_list` with the `Storable` data if we
haven't already fetched. Otherwise doing this is redundant because
`fetch_users_list` sets `@users_list`.

When we do `retrieve`, do the array trick, because that's the only time
when we need to reach inside an outer array. In the loop, just want to
use the plain `@users_list` so that we can have a consistent interface
for both the API-fetched and cache-fetched versions.