complete_profile_field: Add custom fields to completion list
In addition to the Slack defaults, add custom fields as completions to
give users an easy way to modify those as well.

We move the requests for the custom fields (profile endpoint) and
presence information into `find_user` instead of `swhois` because
otherwise the custom fields aren't available to `complete_profile_field`
or `update_user_profile` unless an `/swhois` has previously been run for
the current nick. That was a bug.