find_user: Only request custom fields & presence if not already there
Previously we would make API requests for user profile information and
presence information every time we invoked `find_user($nick)`.

Since we store all users in memory in `@users_list`, once that data is
requested the first time, it's saved for future use.

If the data is already there locally in memory, there's no point in
wasting time and network to request the data again. So only make the API
requests if we don't already have the data.