complete_profile_field: Correctly get current nick
Previously we tried getting the current nick as `$server->{'nick'}`.
This was, however, just a placeholder. Sure, it's how we got the current
nick in an `Irssi::command_bind`, but we don't have a `$server` argument
passed to us in the `complete word` signal.

Instead, we get the current server from the active window passed to the
signal, and from that server then obtain the current nick.

This allows us to add custom field labels to the completion list.

We also wrap all of this in an `if` statement that checks whether there
is an active server on the window object. If not, then we don't add
custom fields to the completion list. Without this condition, trying to
use `$window->{'active_server'}` would prevent _all_ completions from
working. Instead, we still want default Slack profile fields to continue
working because they don't depend on an API call. We just want to
disable completions for custom fields because they depend on a request
to the user profile API endpoint.