print_whois: Pad labels to vertically align colons
Previously, if field labels were longer than five charaters, they
wouldn't be aligned with the colons `:` of the other fields in WHOIS
output.

The lack of alignment didn't look good.

This change pads the field labels with spaces on their right side to
make the colons line up visually and vertically.

We also make a more fundamental change to how the subroutine works.
Instead of trying to print fields immediately using `maybe_print_field`,
they are instead added to an array of hashes. When we want to print
fields, we loop over this array and only print those fields that contain
non-blank values.