swhois: Trim leading & trailing whitespace from nick argument
When the nick is tab completed, it will get a space appended to it which
is included in the `$username` argument passed to our command.

That caused `find_user` to fail because it does an exact string equal
match on the username.

Ensure surrounding spaces are trimmed so that the comparison can be
executed correctly.